requirements.txt
Flask==2.0.1
hvac==1.0.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
Werkzeug==2.0.3